[Changed]Changed the compile option of float size from 4byte to 8byte.
the Azure-sdk-for-C code expects the double size to 8byte.
The detail code that required the 8byte size of double is below Azure-sdk-for-C.

https://github.com/Azure/embedded-wireless-framework/blob/f0a1c89f1ac1fc2182b4005c4bd1f9dc922fa6dd/examples/CK_RX65N_GCCRX/ewf_netx_duo_azure_sdk_pnp_temp_controller_ryz014/libs/netxduo_addons/addons/azure_iot/azure-sdk-for-c/sdk/src/azure/core/az_span_private.h#L54

AZ_NODISCARD AZ_INLINE bool _az_isfinite(double value)

I only change the setting of e2 studio of setting for the compiler.
I did enable the below red square option to set 8byte size of double size.